Refactor shared CSRF checks
